Move Keys to common.ts
